tests: move block-matrix.cpp test to core tests, split up gar test util
